Exercicio 7, feito
