--- exercises/higher-order-functions-do-javascript-es6_3/teste.js
@@ -1,84 +1,0 @@
-const books = [
-    {
-      id: 1,
-      name: 'As Crônicas de Gelo e Fogo',
-      genre: 'Fantasia',
-      author: {
-        name: 'George R. R. Martin',
-        birthYear: 1948,
-      },
-      releaseYear: 1991,
-    },
-    {
-      id: 2,
-      name: 'O Senhor dos Anéis',
-      genre: 'Fantasia',
-      author: {
-        name: 'J. R. R. Tolkien',
-        birthYear: 1892,
-      },
-      releaseYear: 1954,
-    },
-    {
-      id: 3,
-      name: 'Fundação',
-      genre: 'Ficção Científica',
-      author: {
-        name: 'Isaac Asimov',
-        birthYear: 1920,
-      },
-      releaseYear: 1951,
-    },
-    {
-      id: 4,
-      name: 'Duna',
-      genre: 'Ficção Científica',
-      author: {
-        name: 'Frank Herbert',
-        birthYear: 1920,
-      },
-      releaseYear: 1965,
-    },
-    {
-      id: 5,
-      name: 'A Coisa',
-      genre: 'Terror',
-      author: {
-        name: 'Stephen King',
-        birthYear: 1947,
-      },
-      releaseYear: 1986,
-    },
-    {
-      id: 6,
-      name: 'O Chamado de Cthulhu',
-      genre: 'Terror',
-      author: {
-        name: 'H. P. Lovecraft',
-        birthYear: 1890,
-      },
-      releaseYear: 1928,
-    },
-];
-
-
-function oldBooks() {
-  // escreva seu código aqui
-  const year = 2021; 
-  return books 
-  .filter((book) => (year - book.releaseYear) > 60)
-  .map((book) => book.name );
-}
-
-
-console.log(oldBooks());
-
-/* function fantasyOrScienceFictionAuthors() {
-  // escreva seu código aqui
-  const genreBooks = ['Ficção Científica', 'Fantasia']; 
-  return books
-  .filter((book) => genreBooks.includes(book.genre))
-  .map((book) => book.author.name);
- 
-
-} */
